update with new layouts (new Template)

+++ lib/l5m-tools/application.rb
@@ -3,23 +3,45 @@ require 'pathname'
 
 module L5MTools    
     TEMPLATE_DIR = Pathname.new(__FILE__).parent + 'template'    
+    
     class Application
         require 'l5m-tools/tools'
         include Tools        
         def duplicate_app(*args)
             delete = args.delete('-d')
             replacements  = {args[1] => args[2]}
-            File.open(args[0], "r") do |infile|
-                while (line = infile.gets)
-                    line = line.chomp.strip
-                    if line.length > 0 && line[0] != '#'
-                        #duplicate_and_replace( line.chomp ,  replacements ) 
-                        #FileUtils.rm(line, :force => true) if delete
-                        FileUtils.rm(line, :force => true) if duplicate_and_replace( line.chomp ,  replacements )[1] != line && delete
-                    end 
+
+            unless File.exist?(args[0])
+                duplicate_wo_csv(*args)
+            else
+                File.open(args[0], "r") do |infile|
+                    while (line = infile.gets)
+                        line = line.chomp.strip
+                        if line.length > 0 && line[0] != '#'
+                            #duplicate_and_replace( line.chomp ,  replacements ) 
+                            #FileUtils.rm(line, :force => true) if delete
+                            FileUtils.rm(line, :force => true) if duplicate_and_replace( line.chomp ,  replacements )[1] != line && delete
+                        end 
+                    end
                 end
             end
         end
+
+        def duplicate_wo_csv(*args)
+            delete = args.delete('-d')
+            package = args[0]
+            application = args[1]
+            replacements  = {args[1] => args[2]}
+            Dir.glob("#{WORKSPACE}/#{package}/{src,WebContent}/**/#{application}*.*", File::FNM_CASEFOLD).each{|line|  
+                bn = File.basename(line, ".*")
+                #puts line #File.basename(line, ".*") 
+                unless(bn[application.length] =~ /[a-z0-9]/)
+                    #puts line #File.basename(line, ".*") 
+                    FileUtils.rm(line, :force => true) if duplicate_and_replace( line.chomp ,  replacements )[1] != line && delete
+                end                
+            }
+        end
+
         #read file list from file and delete them
         def del(*args)
             replacements  = {args[1] => args[2]}
@@ -42,6 +64,10 @@ module L5MTools
                 "#{WORKSPACE}/#{package}/src/com/l5m/#{package}/engine/worker/#{application}Worker.java"  , replacements ) 
             copy_with_replace( TEMPLATE_DIR+"TemplateServicerImpl.java" , 
                     "#{WORKSPACE}/#{package}/src/com/l5m/#{package}/engine/servicer/#{application}ServicerImpl.java"  , replacements ) 
+
+            copy_with_replace( TEMPLATE_DIR+"TemplateViewer.java" , 
+                    "#{WORKSPACE}/#{package}/src/com/l5m/#{package}/engine/exporter/#{application}Viewer.java"  , replacements ) 
+
             if File.exist?( TEMPLATE_DIR+"#{package}.jsp")
                 copy_with_replace( TEMPLATE_DIR+"#{package}.jsp" ,
                     "#{WORKSPACE}/#{package}/src/jsp/#{application}.jsp" , replacements ) 
@@ -51,8 +77,5 @@ module L5MTools
             end            
             block.call(Time.now, package, application, use_base_worker) if block_given?
         end
-        
-       
-
     end
 end
